Changed clue_touch_t index to size_t type Fixed CLUE_LOG to support logging without interpolation Minor stylistic changes and header reordering

+++ h/clue/touch.h
@@ -1,15 +1,15 @@
 #ifndef CLUE_TOUCH_H
 #define CLUE_TOUCH_H
 
-#include <stddef.h>
-
 #include "clue/internal.h"
+
+#include <stdlib.h>
 
 CLUE_BEGIN_C
 
 typedef struct clue_touch
 {
-	uint8_t index;
+	size_t index;
 	float x;
 	float y;
 	float force;
